# swagger-api-2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "shop-food-api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/wh/unit/create": {
      "post": {
        "tags": [
          "UnitWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/unit/update": {
      "post": {
        "tags": [
          "UnitWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/unit/delete": {
      "post": {
        "tags": [
          "UnitWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/unit/list": {
      "post": {
        "tags": [
          "UnitWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UnitWhListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitWhListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/create": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/update": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/delete": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/list": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CallingConventions": {
        "enum": [
          1,
          2,
          3,
          32,
          64
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CustomAttributeData": {
        "type": "object",
        "properties": {
          "attributeType": {
            "$ref": "#/components/schemas/Type"
          },
          "constructor": {
            "$ref": "#/components/schemas/ConstructorInfo"
          },
          "constructorArguments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomAttributeTypedArgument"
            },
            "nullable": true,
            "readOnly": true
          },
          "namedArguments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomAttributeNamedArgument"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomAttributeTypedArgument": {
        "type": "object",
        "properties": {
          "argumentType": {
            "$ref": "#/components/schemas/Type"
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Exception": {
        "type": "object",
        "properties": {
          "targetSite": {
            "$ref": "#/components/schemas/MethodBase"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "data": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "innerException": {
            "$ref": "#/components/schemas/Exception"
          },
          "helpLink": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "hResult": {
            "type": "integer",
            "format": "int32"
          },
          "stackTrace": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "IntPtr": {
        "type": "object",
        "additionalProperties": false
      },
...
      "MemberTypes": {
        "enum": [
          1,
          2,
          4,
          8,
          16,
          32,
          64,
          128,
          191
        ],
        "type": "integer",
        "format": "int32"
      },
      "MetaData": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exceptionExtra": {
            "$ref": "#/components/schemas/Exception"
          }
        },
        "additionalProperties": false
      },
      "MethodAttributes": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          16,
          32,
          64,
          128,
          256,
          512,
          1024,
          2048,
          4096,
          8192,
          16384,
          32768,
          53248
        ],
        "type": "integer",
        "format": "int32"
      },
      "MethodBase": {
        "type": "object",
        "properties": {
          "memberType": {
            "$ref": "#/components/schemas/MemberTypes"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "declaringType": {
            "$ref": "#/components/schemas/Type"
          },
          "reflectedType": {
            "$ref": "#/components/schemas/Type"
          },
          "module": {
            "$ref": "#/components/schemas/Module"
          },
          "customAttributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomAttributeData"
            },
            "nullable": true,
            "readOnly": true
          },
          "isCollectible": {
            "type": "boolean",
            "readOnly": true
          },
          "metadataToken": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "attributes": {
            "$ref": "#/components/schemas/MethodAttributes"
          },
          "methodImplementationFlags": {
            "$ref": "#/components/schemas/MethodImplAttributes"
          },
          "callingConvention": {
            "$ref": "#/components/schemas/CallingConventions"
          },
          "isAbstract": {
            "type": "boolean",
            "readOnly": true
          },
          "isConstructor": {
            "type": "boolean",
            "readOnly": true
          },
          "isFinal": {
            "type": "boolean",
            "readOnly": true
          },
          "isHideBySig": {
            "type": "boolean",
            "readOnly": true
          },
          "isSpecialName": {
            "type": "boolean",
            "readOnly": true
          },
          "isStatic": {
            "type": "boolean",
            "readOnly": true
          },
          "isVirtual": {
            "type": "boolean",
            "readOnly": true
          },
          "isAssembly": {
            "type": "boolean",
            "readOnly": true
          },
          "isFamily": {
            "type": "boolean",
            "readOnly": true
          },
          "isFamilyAndAssembly": {
            "type": "boolean",
            "readOnly": true
          },
          "isFamilyOrAssembly": {
            "type": "boolean",
            "readOnly": true
          },
          "isPrivate": {
            "type": "boolean",
            "readOnly": true
          },
          "isPublic": {
            "type": "boolean",
            "readOnly": true
          },
          "isConstructedGenericMethod": {
            "type": "boolean",
            "readOnly": true
          },
          "isGenericMethod": {
            "type": "boolean",
            "readOnly": true
          },
          "isGenericMethodDefinition": {
            "type": "boolean",
            "readOnly": true
          },
          "containsGenericParameters": {
            "type": "boolean",
            "readOnly": true
          },
          "methodHandle": {
            "$ref": "#/components/schemas/RuntimeMethodHandle"
          },
          "isSecurityCritical": {
            "type": "boolean",
            "readOnly": true
          },
          "isSecuritySafeCritical": {
            "type": "boolean",
            "readOnly": true
          },
          "isSecurityTransparent": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MethodImplAttributes": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          8,
          16,
          32,
          64,
          128,
          256,
          512,
          4096,
          65535
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Module": {
        "type": "object",
        "properties": {
          "assembly": {
            "$ref": "#/components/schemas/Assembly"
          },
          "fullyQualifiedName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "mdStreamVersion": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "moduleVersionId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "scopeName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "moduleHandle": {
            "$ref": "#/components/schemas/ModuleHandle"
          },
          "customAttributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomAttributeData"
            },
            "nullable": true,
            "readOnly": true
          },
          "metadataToken": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ModuleHandle": {
        "type": "object",
        "properties": {
          "mdStreamVersion": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeMethodHandle": {
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/IntPtr"
          }
        },
        "additionalProperties": false
      },
...
      "UnitWhCreateModelReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnitWhCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "UnitWhCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitWhCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "UnitWhDeleteModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "UnitWhDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "UnitWhDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitWhDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "UnitWhListModelReq": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UnitWhListModelRes": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitWhModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnitWhListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitWhListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "UnitWhModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnitWhUpdateModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnitWhUpdateModelRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnitWhUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitWhUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
...
      "WarehouseCreateModelReq": {
        "required": [
          "address",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "WarehouseCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelReq": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelReq": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelRes": {
        "type": "object",
        "properties": {
          "listWarehouse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarehouseInfoModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}